System: Enable and use c++20
Enable c++20 in top level cmake. Keep c++17 for GrpcInterface, as the protoc tool does not work with c++20.
Several code adjustments to fix compiler issues.

+++ ApplicationLibCode/ProjectDataModel/Summary/RimSummaryCaseCollection.cpp
@@ -72,7 +72,7 @@ void RimSummaryCaseCollection::sortByBinnedVariation( std::vector<RigEnsemblePar
         return;
     }
 
-    double delta = ( maxStdDev - minStdDev ) / RigEnsembleParameter::NR_OF_VARIATION_BINS;
+    double delta = ( maxStdDev - minStdDev ) / (float)( RigEnsembleParameter::NR_OF_VARIATION_BINS );
 
     std::vector<double> bins;
     bins.push_back( 0.0 );
